requirements.txt
absl-py==2.2.2
accelerate==1.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
blis==1.0.1
Bottleneck==1.3.7
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.19.0
comm==0.2.2
confection==0.1.5
cryptography==43.0.3
cymem==2.0.8
datasets==3.1.0
date-spacy==0.0.1
datefinder==0.7.3
dateparser==1.2.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
etils==1.12.2
exceptiongroup==1.2.2
executing==2.1.0
falcon==2.0.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
google==3.0.0
google-auth==2.39.0
google-genai==1.11.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
hug==2.6.1
huggingface-hub==0.26.2
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
joblib==1.4.2
jsonpath-rw==1.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langcodes==3.4.1
language_data==1.2.0
litellm==1.65.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mlcroissant==1.0.17
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.10
nest_asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numexpr==2.8.7
numpy==2.0.2
openai==1.69.0
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pandas-stubs==2.2.3.250308
parso==0.8.4
patsy==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
platformdirs==4.3.6
ply==3.11
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==5.9.0
ptyprocess==0.7.0
pure_eval==0.2.3
py4j==0.10.9.7
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMuPDF==1.24.11
pymupdf4llm==0.0.17
pyparsing==3.2.3
pypdfium2==4.30.0
PySocks==1.7.1
pyspark==3.5.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
pyzmq==25.1.2
quantities==0.16.0
RapidFuzz==3.12.2
rdflib==7.1.4
referencing==0.36.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
rpds-py==0.24.0
rsa==4.9.1
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
selenium==4.25.0
semantic==1.0.3
sentence-transformers==3.3.1
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spark-nlp==5.5.0
srsly==2.4.8
stack-data==0.6.2
statsmodels==0.14.4
sympy==1.13.1
tenacity==9.1.2
thinc==8.3.2
threadpoolctl==3.5.0
tiktoken==0.9.0
torch==2.5.0
torchvision==0.20.0 
torchaudio==2.5.0
--extra-index-url https://download.pytorch.org/whl/cu118
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.48.0
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
types-pytz==2025.2.0.20250326
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.3
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.18.0
zipp==3.21.0